# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: ANDOS API
  version: 1.1.0
  description: AI-Native DevOps System API
servers:
  - url: https://api.andos.dev/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development
security:
  - BearerAuth: []
paths:
...
  /assets/{assetId}/dependencies:
    get:
      summary: Get dependencies
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
            enum: [upstream, downstream]
          description: Dependency direction
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dependency'

    post:
      summary: Create dependency
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDependencyRequest'
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Dependency:
      type: object
      required: [id, source_asset_id, target_asset_id]
      properties:
        id:
          type: string
        source_asset_id:
          type: string
        source_version:
          type: string
        target_asset_id:
          type: string
        target_version:
          type: string
        confirmed_at:
          type: string
          format: date-time
        auto_confirmed:
          type: boolean
    CreateDependencyRequest:
      type: object
      required: [target_asset_id]
      properties:
        target_asset_id:
          type: string
        target_version:
          type: string
        source_version:
          type: string
...